exception tratamento

+++ src/main/java/com/ormlean/commerce/services/ProductService.java
@@ -1,17 +1,21 @@
 package com.ormlean.commerce.services;
 
-import java.util.Optional;
-
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.ormlean.commerce.dto.ProductDTO;
 import com.ormlean.commerce.entities.Product;
 import com.ormlean.commerce.repositories.ProductRepository;
+import com.ormlean.commerce.services.exceptions.DataBaseException;
+import com.ormlean.commerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -23,9 +27,10 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Optional<Product> product = productRepository.findById(id);
+		Product product = productRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado."));
 
-		return mapper.map(product.get(), ProductDTO.class);
+		return mapper.map(product, ProductDTO.class);
 	}
 
 	@Transactional(readOnly = true)
@@ -42,20 +47,31 @@ public class ProductService {
 
 		return mapper.map(product, ProductDTO.class);
 	}
-	
-	
+
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product product = productRepository.getReferenceById(id);		
-		convertDtoToEntity(dto, product);		
-		productRepository.save(product);
-		
-		return mapper.map(product, ProductDTO.class);
+		try {
+			Product product = productRepository.getReferenceById(id);
+			convertDtoToEntity(dto, product);
+			productRepository.save(product);
+
+			return mapper.map(product, ProductDTO.class);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Produto não encontrado.");
+		}
+
 	}
-	
-	@Transactional
+
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		productRepository.deleteById(id);		
+		if (!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Produto não encontrado.");
+		}
+		try {
+			productRepository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException("Falha de integridade referencial.");
+		}
 	}
 
 	public Page<ProductDTO> convertToDtoList(Page<Product> products) {
@@ -65,12 +81,12 @@ public class ProductService {
 	public Product convertDtoToObject(ProductDTO dto) {
 		return mapper.map(dto, Product.class);
 	}
-	
-	private void convertDtoToEntity(ProductDTO dto, Product product) {		
+
+	private void convertDtoToEntity(ProductDTO dto, Product product) {
 		product.setName(dto.getName());
 		product.setDescription(dto.getDescription());
 		product.setImgUrl(dto.getImgUrl());
-		product.setPrice(dto.getPrice());		
+		product.setPrice(dto.getPrice());
 	}
 
 }
